Prevent code point collisions
