+++ packages/icons/scripts/font.js
@@ -3,58 +3,12 @@ const svg2ttf = require('svg2ttf');
 const ttf2eot = require('ttf2eot');
 const ttf2woff = require('ttf2woff');
 const ttf2woff2 = require('ttf2woff2');
-const {
-  Ps: punctuationStartCharacters,
-  Pe: punctuationEndCharacters,
-  Pi: initialPunctuationCharacters,
-  Pf: finalPunctuationCharacters,
-} = require('unicode/category');
+
 
 const pkg = require('../package.json');
-const glyphsMapping = require('../glyphsMapping.json');
 const { readFile, createReadableFromString } = require('./files');
+const { nextCharactersFor } = require('./glyphs');
 const { mirrorSvg } = require('./svg');
-
-const startCharacters = [
-  ...Object.values(punctuationStartCharacters).filter(({ mirrored }) => mirrored === 'Y').map(({ symbol }) => symbol),
-  ...Object.values(initialPunctuationCharacters).filter(({ mirrored }) => mirrored === 'Y').map(({ symbol }) => symbol),
-];
-
-const endCharacters = [
-  ...Object.values(punctuationEndCharacters).filter(({ mirrored }) => mirrored === 'Y').map(({ symbol }) => symbol),
-  ...Object.values(finalPunctuationCharacters).filter(({ mirrored }) => mirrored === 'Y').map(({ symbol }) => symbol),
-];
-
-let directionalCounter = 0;
-let neutralCounter = 0;
-
-const nextCharactersFor = (name, type) => {
-  if (glyphsMapping[name]) {
-    return glyphsMapping[name];
-  }
-
-  if (type === 'dir') {
-    const i = directionalCounter++;
-
-    if (startCharacters.length <= i) {
-      throw Error('more directional icons than possible');
-    }
-
-    glyphsMapping[name] = {
-      start: startCharacters[i],
-      end: endCharacters[i],
-    };
-  }
-
-  if (!type) {
-    const i = neutralCounter++;
-    glyphsMapping[name] = {
-      start: String.fromCodePoint(0xe000 + i),
-    };
-  }
-
-  return glyphsMapping[name];
-};
 
 const createSvgBuffer = async (icons) => {
   const fontStream = new SVGIcons2SVGFontStream({
@@ -119,4 +73,3 @@ module.exports.createTtfBuffer = createTtfBuffer;
 module.exports.createWoffBuffer = createWoffBuffer;
 module.exports.createWoff2Buffer = createWoff2Buffer;
 module.exports.createEotBuffer = createEotBuffer;
-module.exports.glyphsMapping = glyphsMapping;
